Make token objects immutable
@jsphmrrw pointed out that while most of this code is fairly functional,
the token object itself isn't: it's got its own methods and it can
be mutated. As it currently stands, it's a fairly direct translation
from the original implementation in Ruby.

The methods could be pulled out into standalone functions, but I think
the mutability is the biggest problem here. So let's use
`Object.freeze` [1] to prevent the properties of a token from being
changed once it's been created.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/freeze