Implement textParser
In order to parse a tokenized string, we'll need a parser that takes
the output of `tokenize` as input.

Ultimately this will probably be via a `parse` function, but we'll
need a selection of individual parsers that convert tokens (or
collections of tokens) to nodes in an Abstract Syntax Tree.

We could start with the body parser that creates the node at the "top"
of the tree, but I think it'll be easier to implement each of the
lower-level parsers in turn and then use them to build the tree.

So let's start by implementing a `textParser` that can return a text
node or a null node.

I'm not sure whether null nodes really need to be implemented as nodes,
or whether they can just be `false`. For now though I'm trying to stay
as close as possible to the implementation in the article.