Create "matches" function
The text parser only matches a single token, so we didn't need this.

However, subsequent parsers will be looking for patterns consisting of
multiple tokens.

Let's create a "matches" function that tests an array of tokens against
an input pattern. It will return true if the input pattern matches the
types of the first n tokens in the array. The tokens array can be longer
than the pattern: in that case it will only compare the initial tokens.

"matches" should allow us to match more complex chains of tokens in
subsequent parser functions.