Implement "matchStar"
"matchStar" takes a parser and uses it to match against input tokens as
many times as it can.

* If there are 0 matches it returns an empty array
* If there are 1 or more matches, it returns an array containing the
  matched nodes

In the original (Ruby) implementation, matchStar also keeps track of
how many tokens it has consumed. However, doing this made it difficult
to use the same recursive strategy that we already used for
`tokenize` [1]. I'm not sure it really needs to do this: since it
returns an array of nodes, and each node has a `consumed` property,
the place that calls it can also calculate how many tokens
were consumed.

[1] 8ecef24