Implement sentenceAndNewLinesParser
The sentenceAndNewLinesParser looks for any number of sentence fragments
(ie. text, bold, or emphasis) followed by two newline characters.

In order to calculate the number of tokens consumed by the paragraph
parser we need to sum the "consumed" value of all the nodes matched by
the sentence parser, and add two (for the two newline tokens).

In the original (Ruby) implementation [1] this "consumed" value was
returned by "matchStar", but because of the way I implemented that I've
opted to calculate it in the parser (via `Array.reduce`) instead.

[1] https://blog.beezwax.net/2017/08/10/writing-a-markdown-compiler-part-2/