Implement bodyParser
Since we also need to count consumed tokens here (as well as in the
paragraphParser), let's move that `Array.reduce` logic into a standalone
function that will return the total of all "consumed" values, given an
array of nodes.

I'm also adding the paragraphParser here too - I could have separated
this out more tidily, but it's effectively only used by the body parser
to use both sentence and sentenceAndNewLine parsers.

Possibly the paragraphNode is now misleadingly named (because it's
created by the sentenceAndNewLineParser)...