Refactor `matches` function
Previously we were exploding the provided `pattern` and then comparing
the resulting array with the `tokens` array.

However, instead of comparing two arrays, we can simplify things by
converting the `tokens` to a string, and then doing a simpler comparison
between the two resulting strings.

nb. since `tokens` can be longer than `pattern` (`matches` will return
true as long as the first tokens match the pattern), we need to slice
resulting string to be the same length as `pattern` before comparing.