Refactor matcher tests so they don't depend on actual parsers
I'd like to separate out the matcher functions from the parser functions.

In order to do this, let's refactor the tests to use dummy parsers
instead of real parsers. This allows the tests to continue to test
the matchers, but means they no longer depend on the implementation
details of any particular parsers.

This should make it easier to split them out in a subsequent commit.