Remove node creation helper functions
I'd thought having a function to create each distinct node type  might
make for more readable code, but in practice each of these functions is
only called once from its corresponding parser.

If we allow the parser to know about what sort of node it's responsible
for, we can remove the helper functions and everything continues to
work fine. Now I look at it, I think this is a lot simpler too.