Convert JS to ES6 module syntax
Previously the JavaScript was written using the CommonJS format
("require", etc). This was because `jest` and node < 12 use this syntax.

But I'd like to start using this JavaScript on a web page, and for that
it will be easier to use ES6 module syntax, which is now widely supported
by browsers (via `script type="module"`)

In order to keep Jest happy, we have to install
"@babel/plugin-transform-modules-commonjs" as a dependency. This allows
Babel to transform ES6 modules to CommonJS for Jest [1]

[1] https://stackoverflow.com/questions/35756479/does-jest-support-es6-import-export