Create a web page with a live markdown compiler
Because this is all written in JavaScript, it's easy enough to import
into a web page and view the data structure that results from running
Markdown through `parse`

You can type in the textarea, and the result will appear below as
you type.

You'll also need a way of serving the page. I like `browser-sync`,
which you can install like this:

    npm install --global browser-sync

...and then run in the current directory like this:

    browser-sync