Allow live markdown compiler to show different output types
Previously we were showing the output of `parse`, which is the result
of parsing tokens into nodes.

But the process of parsing the markdown goes like this:

    Markdown --> Tokens --> Nodes --> HTML

Let's add some toggles so we can see the data structures at each point.

The first toggle shows the output of `tokenize` - ie. an array of tokens.
The second shows the output of `parse` - ie. an abstract syntax tree.

The step that converts the AST to HTML is not yet implemented, but let's
add a third toggle to make it clearer that there is a final step in
the process.