Remove dependency of parser on tokenizer
Previously, `parse` was expecting text input and using `tokenize` to
convert to tokens.

In order to keep the parser and tokenizer self-contained, let's tweak
`parse` so it expects tokens as input. Then we can make whatever calls it
responsible for first tokenizing any input.

This change is reflected in a tweak to how we're parsing the input in
the browser - calling `parse` on the output of `tokenize` (and `render`
on the output of `parse`) makes it clearer how these things interrelate.

The `parse` function is now effectively an alias for `bodyParser` -
this could be refactored but I'll leave that for some other time.
Likewise, it should really have a test.