Create generator
The HTML is generated using the Visitor pattern [1]. The Ruby approach
uses a number of different classes, but in JavaScript I'm thinking it
might be more idiomatic to try using a recursive `visit` function instead.

First off, let's create a simple `visit` function that simply crawls an
Abstract Syntax Tree and returns a concatenated string containing the
values of all the nodes.

Subsequently, we can add rules for handling particular node types.

Let's also update the HTML page so we can try it out in the browser.
Because I'd already called it "render" here, let's just import "generate"
as "render" for now.

[1] https://blog.beezwax.net/2018/05/25/writing-a-markdown-compiler-part-3/